Use ObjectIds instead of String UUIDs using setting
If use_objectids is true, and a doc to be saved does not have an _id, no string UUID is set. As a result, the Java driver will fill in the _id with a new ObjectId. In this case, the new ObjectId is returned instead of a new string UUID, in the format {"$oid":"1234567890AB1234567890AB"}.
use_objectids defaults to false, to preserve old behaviour.
In order to be useful, this also requires use_mongo_types to be true.
There is a new javascript test for saving with use_objectids enabled, which checks that the expected format of ObjectId JSON is received, and uses that id to perform an update and findone, as in the existing testSave test.
Added description of option to README.md
This is based on https://github.com/vert-x/mod-mongo-persistor/pull/41/files but with the ObjectId returned, a test added, and based on the current master of vertx/mod-mongo-persistor.